feat: skip around queue by clicking

+++ frontend/src/message.rs
@@ -19,7 +19,7 @@ pub enum Message {
 
     PlaySong(backend::Song),
     Queue(backend::QueueEvent),
-    Skip(i32),
+    Skip(isize),
     ResumeSong,
     PauseSong,
